remove substack dmoain cache
